● Make LogicalTimestamp DB-serializable
  Move LogicalTimestamp out of types/primitives.rs (the in-memory-only
  module) into its own file under types/db/, where DB-serializable
  primitives belong per the module convention. Add the two on-disk
  contracts the type was missing: ZainoVersionedSerde (V1, 4-byte
  big-endian body, same shape as Height) and FixedEncodedLen (4 bytes).

  Big-endian is the load-bearing choice: it makes lexicographic byte
  order over LMDB keys match numeric logical_ts order, which is the
  property cursor-based range scans depend on. The forthcoming
  logical_ts -> BlockHash index (zingolabs/zaino#1101) needs this; so
  will any migration that rebuilds entries in sorted order.

  The recurrence itself (LogicalTimestamp::next) moves intact alongside
  its struct so the encoding and the algorithm stay together — they
  share the same correctness story.

  Three new tests pin the on-disk contract:
  - encode_v1_is_big_endian_u32: byte-for-byte golden for value 1
  - encode_decode_round_trips: 0, 1, mid-range, u32::MAX
  - encoded_bytes_sort_lexicographically: confirms the LMDB-cursor
    property holds across u8-boundary values (255 vs 256, etc.)

  No call site changes: every consumer goes through the
  chain_index::types::LogicalTimestamp re-export, which now resolves to
  the new path. No DB schema change. Pure type-promotion that unblocks
  every subsequent step toward the logical_ts index.
